deployment/requirements.txt
fastapi
uvicorn
ultralytics
transformers
torch
Pillow
word2number
python-multipart
